etcd周りをmodel, controllerに分離
