+++ api-server/srcs/controllers/related_terms_controller.py
@@ -2,47 +2,10 @@
 from flask import Blueprint, current_app, request, jsonify
 from ..models.related_terms_model import FasttextWrapper
 from ..models.noun_detect_model import GinzaWrapper
-import etcd3
+from ..models.cache_model import EtcdWrapper
 import json
 
 app = Blueprint("synonym", __name__)
-
-# etcd 初期化
-etcd = etcd3.client(host="etcd", port=2379)
-
-
-@app.route("/api/v1/dev/etcd")
-def etcd_test():
-    key = request.args.get("key")
-    response = {}
-    value = etcd.get(key)
-    if cache_exists(key) is True:
-        response = {"key": key, "value": value[0].decode()}
-    else:
-        print("cache doesn't exists")
-        # TODO create cache
-        # etcd.put(key, synonyms)
-
-    return jsonify(response)
-
-
-def get_cache(key):
-    value = etcd.get(key)
-    return value[0].decode()
-
-
-def put_cache(key, value):
-    jsonstring = json.dumps(value, ensure_ascii=False)
-    etcd.put(key, jsonstring)
-
-
-def cache_exists(key):
-    value = etcd.get(key)
-    if value[0] is None:
-        # cache doesn't exist
-        return False
-    else:
-        return True
 
 
 @app.route("/api/v1/synonym")
@@ -57,16 +20,18 @@ def get_synonym():
         sentence = request.args.get("sentence")
 
         # 既にキャッシュに結果が存在すれば使い回して高速化
-        if cache_exists(sentence) is True:
-            return jsonify(json.loads(get_cache(sentence)))
+        etcd = EtcdWrapper()
+        if etcd.cache_exists(sentence):
+            return jsonify(json.loads(etcd.get_cache(sentence)))
         else:
             # GiNZA（SudachiPy）による名詞抽出
             noun_list = GinzaWrapper().get_noun(sentence)
+            # fastTextで意味の近い単語を抽出
             get_synonym_core(noun_list, response)
             logger.info("finish synonym")
 
             # キャッシュに追加
-            put_cache(sentence, response)
+            etcd.put_cache(sentence, response)
 
             return jsonify(response)
 
@@ -78,6 +43,9 @@ def get_synonym():
 
 # TODO 高速化
 def get_synonym_core(noun_list, response):
+    """
+    fastTextで意味の近い単語を抽出
+    """
     ft = FasttextWrapper()
     for noun in noun_list:
         similar_words = ft.get_similar_words(noun.lower())
